Extract title tooltip
